Revise libsingular_julia
Adapt to latest CxxWrap; in particular, build separate binaries for each
minor Julia version (1.3, 1.4, 1.5, ...)

This also adds support for more platforms.

The build dependencies on GMP_jll and MPFR_jll were only implicit so far,
but we need to make them explicit and with version constraints to avoid
issues on macOS (when this JLL is built against a too-new version of either
of these libraries, it becomes unloadable in older Julia versions)
