+++ L/libsingular_julia/libsingular_julia@1.3/build_tarballs.jl
@@ -1,0 +1,2 @@
+const julia_version = v"1.3.1"
+include("../common.jl")
